Add governance flag to pause the bridge
This commit adds a governance flag that can be used to temporarily halt
the movement of tokens across the bridge during emergencies.

This flag is under the control of governance making it somewhat slow for
an emergency halt mechanism, but it's superior to a centralized bridge
killswitch which is the only other alternative.

Pausing the bridge does have a few gotcha's namely that signature
submissions from validators are still required, becuase paused or not
the validator set on Ethereum must be kept in sync with Comsos and that
chain state will grow more quickly while the bridge is puased, becuase
validator set updates and attestations are not being pruned.