Use hashmap for slashing confirms
This patch flattens the innermost loop of our slashing conditions by
building a hashmap of confirms at the start of the loop and then simply
performing a membership check. Since we're not iterating over the
hashmap this will introduce no non-determinism.

As a secondary effect of this change we are now consistently setting
orchestrator addresses in our unit test environment rather than using
ethereum addresses to look up the validator in question.