Move burn/mint from tx pool to batch execution
Burning Ethereum-originated assets on tx pool entry requires re-minting
those assets if a tx is removed from the pool. Instead we now transfer
the assets to the module, wait for the tx to appear in an observed
batch, and burn the tokens then. Minting has been removed from
tx cancellation.