Remove OutgoingTxBatchBlockKey
This key represented a second index of tx batches by block. Not only was
this second index wasteful in terms of storage, since it kept a
duplicate of the entire block. But it also risked being inconsistent
with the main block index.

This is now no longer possible as we simply iterate over outgoing blocks
in general rather than using a larger storage index.