Migrate attestation tests to five val unit test env
As an unintended consequence of changing where the current bonded set is
loaded in the endblocker we can no longer use our dummy panic-only
keeper for these tests.

Fortunately we already have an improved unit testing environment, but
it's not used everywhere, this patch migrates attestation based tests to
this new environment.

This is in general a good improvement, allowing these tests
to interact with a more complete, but still mocked environment.