Add IBC chain to test environment
This patch adds a test chain based off of Cosmos-sdk that also runs
within our testing docker container. Cosmos-sdk was chosen as it is sure
to have a tag for any version gravity itself may ever be on.

Constants for accessing this chain have been added to the test
environment, but currently they are not used.