Fix proto tooling and setup instructions
The proto generator script was not using the correct generator plugin
for service gateways, this worked correctly on some machines that had
the plugin installed in an undocumented way, but not if the
docs where followed to the letter on a new machine.

This patch also includes improvements to the docs. The go install
command for protoc-gen-gocosmos requires a go get for a dependency after
which you must re-run the previous co install command. If you're not
paying close attention and never re-reun the initial command you won't
have the right plugin.

Co-authored-by: Thomas Spellman <thomas@althea.systems>